requirements.txt
discord.py
pymysql==0.9.3
aiomysql==0.0.21
aiohttp
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
pyowm
dblpy
osuapi